Some Menu optimization - has still to be taken care of :-) Added a spectrum analyzer which shows the incomming signal spectrum in the top row and the minimum of Noisereduction coefficients in the 2nd row

+++ NR3-eclipse/Audio_DSP/inc/audio_nr.h
@@ -9,7 +9,7 @@
 
 
 typedef struct {
-  uint16_t Version;
+  int32_t Version;
   int32_t alpha_int;//0..100
   int32_t asnr_int; //0..30
   int32_t power_threshold_int;  //0..100
